api/requirements.txt
azure-functions
openai>=1.0.0
requests 